requirements.txt
Flask
notion-client
pyttsx3
python-dotenv
google-genai
pygame
